Start of a new stone!
merged trinity branch with trunk, trunk version can now support both mangos and trinity,
only thing left out was the template logos, can someone redo it as "Mangos/Trinity"?
trinity branch is now considered defunct.

trinity users, remember to set $server_type = 1; in config

We are in the process of major reorganize of MiniManager code, so we think it is best to merge. We know back porting isn't fun.

trinity users please test this out, and forward your feedback to:
http://mangos.osh.nu/forums/index.php?showforum=19